backend/requirements.txt
flask
flask-jwt-extended
mysql-connector-python
python-dotenv
werkzeug
flask-cors